Refactored serial connetion directory
