+++ src-tauri/src/device/connections/mod.rs
@@ -1,0 +1,1 @@
+pub mod serial;
