+++ src-tauri/src/device/connections/serial/helpers.rs
