+++ src-tauri/src/device/connections/serial/mod.rs
